social_network/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
cffi==1.17.1
cryptography==43.0.1
dj-database-url==2.2.0
Django==5.1.1
django-appconf==1.0.6
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.1
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
redis==5.0.8
sqlparse==0.5.1
typing_extensions==4.12.2
